Desktop/Projects/mail-manager/requirements.txt
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
msal>=1.28.0
requests>=2.31.0
anthropic>=0.49.0
google-cloud-storage>=2.16.0
google-cloud-secret-manager>=2.20.0
functions-framework>=3.5.0
